core: isolate authority storage adapter ( spec001w6 )
Keep controller policy storage-agnostic and let Core mint opaque operation capabilities only after validating non-authorizing redb preparation data.
