+++ packages/d2b-core-controller/Cargo.toml
@@ -20,8 +20,6 @@ d2b-audit = { path = "../d2b-audit", version = "0.0.0-bootstrap" }
 d2b-contracts = { path = "../d2b-contracts", version = "0.0.0-bootstrap" }
 d2b-controller-toolkit = { path = "../d2b-controller-toolkit", version = "0.0.0-bootstrap" }
 d2b-telemetry = { path = "../d2b-telemetry", version = "0.0.0-bootstrap" }
-d2b-resource-store-redb = { path = "../d2b-resource-store-redb", version = "0.0.0-bootstrap" }
-d2b-resource-store = { path = "../d2b-resource-store", version = "0.0.0-bootstrap" }
 serde.workspace = true
 serde_json.workspace = true
 sha2.workspace = true
