+++ packages/d2b-core-controller/src/authority_persistence.rs
@@ -4,20 +4,11 @@
 //! typed row and recovery validation. Only this adapter can turn storage rows
 //! into the private receipt consumed by `HostGlobalAuthorityIndex`.
 
-use std::{
-    collections::BTreeMap,
-    future::Future,
-    pin::Pin,
-    sync::{Arc, Mutex},
-};
+use std::{collections::BTreeMap, future::Future, pin::Pin, sync::Arc};
 
 use crate::authority::{
     AuthorityOperationState, AuthorityRecoveryReceipt, AuthorityStorageClaim,
-    AuthorityStorageOperation, ExternalNicRecoveryInventory, HostGlobalAuthorityIndex,
-};
-use d2b_resource_store::{StoreOperationContext, StoreResolveRequest};
-use d2b_resource_store_redb::{
-    AuthorityOperation, AuthorityOperationState as StoreAuthorityOperationState, RedbResourceStore,
+    AuthorityStorageOperation, HostGlobalAuthorityIndex,
 };
 
 /// Stable persistence error without host paths or storage details.
@@ -46,8 +37,72 @@ impl std::error::Error for AuthorityPersistenceError {}
 pub type AuthorityFuture<'a, T> =
     Pin<Box<dyn Future<Output = Result<T, AuthorityPersistenceError>> + Send + 'a>>;
 
-mod sealed {
-    pub trait RecoveryProvenance {}
+/// Non-authorizing evidence returned by a storage adapter after it has
+/// durably prepared one operation.
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct PreparedAuthorityOperation {
+    operation_id: String,
+    store_binding_digest: String,
+    nonce: u64,
+}
+
+impl PreparedAuthorityOperation {
+    pub fn new(
+        operation_id: String,
+        store_binding_digest: String,
+        nonce: u64,
+    ) -> Result<Self, AuthorityPersistenceError> {
+        if operation_id.is_empty()
+            || operation_id.len() > 512
+            || store_binding_digest.is_empty()
+            || nonce == 0
+        {
+            return Err(AuthorityPersistenceError::RowInvalid);
+        }
+        Ok(Self {
+            operation_id,
+            store_binding_digest,
+            nonce,
+        })
+    }
+
+    pub(crate) fn matches_operation(&self, operation: &AuthorityStorageOperation) -> bool {
+        self.operation_id == operation.operation_id
+            && self.store_binding_digest == operation.store_binding_digest
+            && self.nonce != 0
+    }
+}
+
+/// Non-authorizing recovery data returned by a storage adapter. Core validates
+/// the operations and provenance before minting any authority capabilities.
+pub struct AuthorityRecoveryData {
+    operations: Vec<AuthorityStorageOperation>,
+    prepared_operations: BTreeMap<String, PreparedAuthorityOperation>,
+}
+
+impl AuthorityRecoveryData {
+    pub fn new(
+        operations: Vec<AuthorityStorageOperation>,
+        prepared_operations: BTreeMap<String, PreparedAuthorityOperation>,
+    ) -> Self {
+        Self {
+            operations,
+            prepared_operations,
+        }
+    }
+
+    pub(crate) fn into_parts(
+        self,
+    ) -> (
+        Vec<AuthorityStorageOperation>,
+        BTreeMap<String, PreparedAuthorityOperation>,
+    ) {
+        (self.operations, self.prepared_operations)
+    }
+
+    pub(crate) fn operations(&self) -> &[AuthorityStorageOperation] {
+        &self.operations
+    }
 }
 
 /// Opaque Core-owned handle for one prepared operation.
@@ -55,6 +110,25 @@ mod sealed {
 /// The operation id is intentionally not exposed as a public accessor. A
 /// persistence implementation may bind this handle to its store instance and
 /// reject handles minted by another store.
+///
+/// ```compile_fail
+/// use d2b_core_controller::authority_persistence::AuthorityOperationCapability;
+/// let _ = AuthorityOperationCapability::new(
+///     "operation".to_owned(),
+///     "sha256:".to_owned() + &"0".repeat(64),
+///     1,
+/// );
+/// ```
+///
+/// ```compile_fail
+/// use d2b_core_controller::authority_persistence::AuthorityOperationCapability;
+/// let _ = AuthorityOperationCapability::default();
+/// ```
+///
+/// ```compile_fail
+/// use d2b_core_controller::authority_persistence::AuthorityOperationCapability;
+/// let _: AuthorityOperationCapability = serde_json::from_str("{}").unwrap();
+/// ```
 pub struct AuthorityOperationCapability {
     pub(crate) operation_id: String,
     pub(crate) store_binding_digest: String,
@@ -68,19 +142,33 @@ impl core::fmt::Debug for AuthorityOperationCapability {
 }
 
 impl AuthorityOperationCapability {
-    pub(crate) fn new(
-        operation_id: String,
-        store_binding_digest: String,
-        nonce: u64,
+    pub(crate) fn from_prepared(
+        operation_id: &str,
+        prepared: PreparedAuthorityOperation,
     ) -> Result<Self, AuthorityPersistenceError> {
-        if operation_id.is_empty() || operation_id.len() > 512 || store_binding_digest.is_empty() {
+        if prepared.operation_id != operation_id {
             return Err(AuthorityPersistenceError::RowInvalid);
         }
         Ok(Self {
-            operation_id,
-            store_binding_digest,
-            nonce,
+            operation_id: prepared.operation_id,
+            store_binding_digest: prepared.store_binding_digest,
+            nonce: prepared.nonce,
         })
+    }
+
+    #[doc(hidden)]
+    pub fn operation_id(&self) -> &str {
+        &self.operation_id
+    }
+
+    #[doc(hidden)]
+    pub fn store_binding_digest(&self) -> &str {
+        &self.store_binding_digest
+    }
+
+    #[doc(hidden)]
+    pub const fn nonce(&self) -> u64 {
+        self.nonce
     }
 }
 
@@ -91,7 +179,7 @@ pub trait AuthorityPersistence: Send + Sync {
         &'a self,
         operation_id: &'a str,
         claim: &'a AuthorityStorageClaim,
-    ) -> AuthorityFuture<'a, AuthorityOperationCapability>;
+    ) -> AuthorityFuture<'a, PreparedAuthorityOperation>;
 
     /// Record the effect outcome while retaining ownership.
     fn record_effect<'a>(
@@ -114,261 +202,22 @@ pub trait AuthorityPersistence: Send + Sync {
 
     /// Load every non-terminal row and terminal operation identity before
     /// new admission.
-    fn recover<'a>(&'a self) -> AuthorityFuture<'a, AuthorityRecoveryReceipt>;
+    fn recover<'a>(&'a self) -> AuthorityFuture<'a, AuthorityRecoveryData>;
 }
 
 /// Trusted provenance port for recovery rows. Implementations must validate
 /// authoritative owner generations and external inventory identities; digest
 /// equality alone is not sufficient.
-pub trait AuthorityRecoveryProvenance: sealed::RecoveryProvenance + Send + Sync {
+pub trait AuthorityRecoveryProvenance: Send + Sync {
     fn validate<'a>(&'a self, operation: &'a AuthorityStorageOperation) -> AuthorityFuture<'a, ()>;
-}
-
-/// Production authority persistence owner for one Zone redb store.
-pub struct RedbAuthorityPersistence {
-    store: Arc<RedbResourceStore>,
-    operation_capabilities:
-        Mutex<BTreeMap<String, Arc<d2b_resource_store_redb::AuthorityOperationCapability>>>,
-    external_inventory: Option<Arc<dyn ExternalNicRecoveryInventory>>,
-}
-
-impl core::fmt::Debug for RedbAuthorityPersistence {
-    fn fmt(&self, formatter: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
-        formatter.write_str("RedbAuthorityPersistence(<store-bound>)")
-    }
-}
-
-impl RedbAuthorityPersistence {
-    /// Bind the port to the already opened, broker-owned Zone store.
-    pub fn new(store: Arc<RedbResourceStore>) -> Self {
-        Self {
-            store,
-            operation_capabilities: Mutex::new(BTreeMap::new()),
-            external_inventory: None,
-        }
-    }
-
-    pub fn with_external_inventory(
-        mut self,
-        inventory: Arc<dyn ExternalNicRecoveryInventory>,
-    ) -> Self {
-        self.external_inventory = Some(inventory);
-        self
-    }
-}
-
-impl AuthorityPersistence for RedbAuthorityPersistence {
-    fn prepare<'a>(
-        &'a self,
-        operation_id: &'a str,
-        claim: &'a AuthorityStorageClaim,
-    ) -> AuthorityFuture<'a, AuthorityOperationCapability> {
-        Box::pin(async move {
-            let claim_digest = crate::authority::claim_digest(claim)
-                .map_err(|_| AuthorityPersistenceError::RowInvalid)?;
-            let owner_ref_present = match claim {
-                AuthorityStorageClaim::Generic(claim) => {
-                    claim.owner_proof().resource_ref().is_some()
-                }
-                AuthorityStorageClaim::ExternalNic(claim) => {
-                    claim.owner_proof().resource_ref().is_some()
-                }
-            };
-            if !owner_ref_present {
-                return Err(AuthorityPersistenceError::RowInvalid);
-            }
-            let binding_digest = self.store.authority_binding_digest(&claim_digest);
-            let row = AuthorityStorageOperation {
-                operation_id: operation_id.to_owned(),
-                claim: claim.clone(),
-                state: AuthorityOperationState::Pending,
-                claim_digest: claim_digest.clone(),
-                store_binding_digest: binding_digest.clone(),
-            };
-            AuthorityRecoveryProvenance::validate(self, &row).await?;
-            let payload =
-                serde_json::to_vec(&row).map_err(|_| AuthorityPersistenceError::RowInvalid)?;
-            let capability = self
-                .store
-                .prepare_authority_operation(operation_id.to_owned(), payload, &claim_digest)
-                .await
-                .map_err(|_| AuthorityPersistenceError::CommitUnknown)?;
-            let nonce = capability.nonce();
-            self.operation_capabilities
-                .lock()
-                .map_err(|_| AuthorityPersistenceError::StateInvalid)?
-                .insert(operation_id.to_owned(), Arc::new(capability));
-            AuthorityOperationCapability::new(operation_id.to_owned(), binding_digest, nonce)
-        })
-    }
-
-    fn record_effect<'a>(
-        &'a self,
-        capability: &'a AuthorityOperationCapability,
-        state: AuthorityOperationState,
-    ) -> AuthorityFuture<'a, ()> {
-        Box::pin(async move {
-            let state = match state {
-                AuthorityOperationState::Pending => StoreAuthorityOperationState::Pending,
-                AuthorityOperationState::EffectConfirmed => {
-                    StoreAuthorityOperationState::EffectConfirmed
-                }
-                AuthorityOperationState::EffectRetryable => {
-                    StoreAuthorityOperationState::EffectRetryable
-                }
-                AuthorityOperationState::EffectTerminal => {
-                    StoreAuthorityOperationState::EffectTerminal
-                }
-                AuthorityOperationState::Closing => StoreAuthorityOperationState::Closing,
-                AuthorityOperationState::Closed | AuthorityOperationState::Released => {
-                    return Err(AuthorityPersistenceError::StateInvalid);
-                }
-            };
-            let store_capability = self.validate_capability(capability)?;
-            store_capability
-                .record_effect(state)
-                .await
-                .map_err(|_| AuthorityPersistenceError::StoreUnavailable)
-        })
-    }
-
-    fn record_close<'a>(
-        &'a self,
-        capability: &'a AuthorityOperationCapability,
-    ) -> AuthorityFuture<'a, ()> {
-        Box::pin(async move {
-            let store_capability = self.validate_capability(capability)?;
-            store_capability
-                .record_close()
-                .await
-                .map_err(|_| AuthorityPersistenceError::StoreUnavailable)
-        })
-    }
-
-    fn release<'a>(
-        &'a self,
-        capability: &'a AuthorityOperationCapability,
-    ) -> AuthorityFuture<'a, ()> {
-        Box::pin(async move {
-            let store_capability = self.validate_capability(capability)?;
-            store_capability
-                .release()
-                .await
-                .map_err(|_| AuthorityPersistenceError::StoreUnavailable)?;
-            self.operation_capabilities
-                .lock()
-                .map_err(|_| AuthorityPersistenceError::StateInvalid)?
-                .remove(&capability.operation_id);
-            Ok(())
-        })
-    }
-
-    fn recover<'a>(&'a self) -> AuthorityFuture<'a, AuthorityRecoveryReceipt> {
-        Box::pin(async move {
-            let rows = self
-                .store
-                .authority_operations()
-                .await
-                .map_err(|_| AuthorityPersistenceError::StoreUnavailable)?;
-            recovery_receipt(rows, &self.store, &self.operation_capabilities).await
-        })
-    }
-}
-
-impl sealed::RecoveryProvenance for RedbAuthorityPersistence {}
-
-impl AuthorityRecoveryProvenance for RedbAuthorityPersistence {
-    fn validate<'a>(&'a self, operation: &'a AuthorityStorageOperation) -> AuthorityFuture<'a, ()> {
-        Box::pin(async move {
-            let claim_digest = crate::authority::claim_digest(&operation.claim)
-                .map_err(|_| AuthorityPersistenceError::RowInvalid)?;
-            if claim_digest != operation.claim_digest
-                || self.store.authority_binding_digest(&claim_digest)
-                    != operation.store_binding_digest
-            {
-                return Err(AuthorityPersistenceError::RowInvalid);
-            }
-            if !matches!(
-                operation.state,
-                AuthorityOperationState::Closed | AuthorityOperationState::Released
-            ) {
-                let owner_proof = match &operation.claim {
-                    AuthorityStorageClaim::Generic(claim) => claim.owner_proof(),
-                    AuthorityStorageClaim::ExternalNic(claim) => claim.owner_proof(),
-                };
-                let Some(owner_ref) = owner_proof.resource_ref() else {
-                    return Err(AuthorityPersistenceError::RowInvalid);
-                };
-                let resolved = self
-                    .store
-                    .resolve_ref(StoreResolveRequest {
-                        operation: StoreOperationContext {
-                            operation_id: format!("authority-recovery:{}", operation.operation_id),
-                            idempotency_key: None,
-                            correlation_id: "authority-recovery".to_owned(),
-                            trace_id: None,
-                            deadline_ms: 1,
-                        },
-                        zone: self.store.identity().zone().clone(),
-                        target: owner_ref.clone(),
-                        expected_uid: Some(owner_proof.resource_uid().clone()),
-                    })
-                    .await
-                    .map_err(|_| AuthorityPersistenceError::RowInvalid)?;
-                if resolved.uid != *owner_proof.resource_uid()
-                    || resolved.generation != owner_proof.generation()
-                {
-                    return Err(AuthorityPersistenceError::RowInvalid);
-                }
-            }
-            if !matches!(
-                operation.state,
-                AuthorityOperationState::Closed | AuthorityOperationState::Released
-            ) && matches!(operation.claim, AuthorityStorageClaim::ExternalNic(_))
-            {
-                let AuthorityStorageClaim::ExternalNic(claim) = &operation.claim else {
-                    unreachable!();
-                };
-                let Some(inventory) = &self.external_inventory else {
-                    return Err(AuthorityPersistenceError::RowInvalid);
-                };
-                if !inventory.contains_identity(claim.host_uid(), claim.identity_digest()) {
-                    return Err(AuthorityPersistenceError::RowInvalid);
-                }
-            }
-            Ok(())
-        })
-    }
-}
-
-impl RedbAuthorityPersistence {
-    fn validate_capability(
-        &self,
-        capability: &AuthorityOperationCapability,
-    ) -> Result<Arc<d2b_resource_store_redb::AuthorityOperationCapability>, AuthorityPersistenceError>
-    {
-        let store_capability = self
-            .operation_capabilities
-            .lock()
-            .map_err(|_| AuthorityPersistenceError::StateInvalid)?
-            .get(&capability.operation_id)
-            .cloned()
-            .ok_or(AuthorityPersistenceError::StateInvalid)?;
-        if capability.store_binding_digest.is_empty()
-            || capability.nonce != store_capability.nonce()
-            || !store_capability.matches_binding_digest(&capability.store_binding_digest)
-        {
-            return Err(AuthorityPersistenceError::StateInvalid);
-        }
-        Ok(store_capability)
-    }
 }
 
 /// Rehydrate the central index from the production Zone persistence owner.
 pub async fn rehydrate_from_persistence(
     persistence: &dyn AuthorityPersistence,
+    provenance: &dyn AuthorityRecoveryProvenance,
 ) -> Result<HostGlobalAuthorityIndex, AuthorityPersistenceError> {
-    let receipt = persistence.recover().await?;
+    let receipt = validated_recovery_receipt(persistence.recover().await?, provenance).await?;
     HostGlobalAuthorityIndex::rehydrate(receipt).map_err(|_| AuthorityPersistenceError::RowInvalid)
 }
 
@@ -391,10 +240,7 @@ impl AuthorityRecoveryCoordinator {
         persistence: Arc<dyn AuthorityPersistence>,
         provenance: &dyn AuthorityRecoveryProvenance,
     ) -> Result<Self, AuthorityPersistenceError> {
-        let receipt = persistence.recover().await?;
-        for operation in receipt.operations() {
-            provenance.validate(operation).await?;
-        }
+        let receipt = validated_recovery_receipt(persistence.recover().await?, provenance).await?;
         let index = HostGlobalAuthorityIndex::rehydrate(receipt)
             .map_err(|_| AuthorityPersistenceError::RowInvalid)?;
         Ok(Self {
@@ -464,76 +310,18 @@ impl AuthorityRecoveryCoordinator {
     }
 }
 
-async fn recovery_receipt(
-    rows: Vec<AuthorityOperation>,
-    store: &Arc<RedbResourceStore>,
-    operation_capabilities: &Mutex<
-        BTreeMap<String, Arc<d2b_resource_store_redb::AuthorityOperationCapability>>,
-    >,
+async fn validated_recovery_receipt(
+    data: AuthorityRecoveryData,
+    provenance: &dyn AuthorityRecoveryProvenance,
 ) -> Result<AuthorityRecoveryReceipt, AuthorityPersistenceError> {
-    let mut operations = Vec::new();
-    let mut capabilities = BTreeMap::new();
-    let mut operation_ids = std::collections::BTreeSet::new();
-
-    for row in rows {
-        if !operation_ids.insert(row.operation_id.clone()) {
-            return Err(AuthorityPersistenceError::RowInvalid);
-        }
-        let mut stored: AuthorityStorageOperation = serde_json::from_slice(&row.payload)
-            .map_err(|_| AuthorityPersistenceError::RowInvalid)?;
-        if stored.operation_id != row.operation_id {
-            return Err(AuthorityPersistenceError::RowInvalid);
-        }
-        if stored.store_binding_digest != store.authority_binding_digest(&stored.claim_digest) {
-            return Err(AuthorityPersistenceError::RowInvalid);
-        }
-        // The redb lifecycle column is the authoritative state. The payload
-        // is an untrusted claim envelope and older physical rows may carry
-        // the prepare-time state, so never let it override the committed
-        // transition.
-        stored.state = match row.state {
-            StoreAuthorityOperationState::Pending => AuthorityOperationState::Pending,
-            StoreAuthorityOperationState::EffectConfirmed => {
-                AuthorityOperationState::EffectConfirmed
-            }
-            StoreAuthorityOperationState::EffectRetryable => {
-                AuthorityOperationState::EffectRetryable
-            }
-            StoreAuthorityOperationState::EffectTerminal => AuthorityOperationState::EffectTerminal,
-            StoreAuthorityOperationState::Closing => AuthorityOperationState::Closing,
-            StoreAuthorityOperationState::Closed => AuthorityOperationState::Closed,
-            StoreAuthorityOperationState::Released => AuthorityOperationState::Released,
-        };
-        if !matches!(
-            stored.state,
-            AuthorityOperationState::Closed | AuthorityOperationState::Released
-        ) {
-            let store_capability = store
-                .resume_authority_operation(
-                    stored.operation_id.clone(),
-                    &stored.store_binding_digest,
-                )
-                .await
-                .map_err(|_| AuthorityPersistenceError::RowInvalid)?;
-            let nonce = store_capability.nonce();
-            operation_capabilities
-                .lock()
-                .map_err(|_| AuthorityPersistenceError::StateInvalid)?
-                .insert(stored.operation_id.clone(), Arc::new(store_capability));
-            let capability = AuthorityOperationCapability::new(
-                stored.operation_id.clone(),
-                stored.store_binding_digest.clone(),
-                nonce,
-            )?;
-            capabilities.insert(stored.operation_id.clone(), capability);
-        }
-        operations.push(stored);
+    for operation in data.operations() {
+        provenance.validate(operation).await?;
     }
-
-    HostGlobalAuthorityIndex::recovery_receipt_from_operations_with_capabilities(
+    let (operations, prepared_operations) = data.into_parts();
+    HostGlobalAuthorityIndex::recovery_receipt_from_operations_with_prepared_capabilities(
         operations,
         None,
-        capabilities,
+        prepared_operations,
     )
     .map_err(|_| AuthorityPersistenceError::RowInvalid)
 }
