+++ tests/golden/api-surface/capability-trait-impls.txt
@@ -884,10 +884,6 @@ d2b_core_controller::d2b_core_controller::authority::emergency_policy::Emergency
 d2b_core_controller::d2b_core_controller::authority::quota::QuotaAuthority	struct	core::core::fmt::Debug	trait	impl core::fmt::Debug for d2b_core_controller::authority::quota::QuotaAuthority
 d2b_core_controller::d2b_core_controller::authority::quota::Usage	struct	core::core::default::Default	trait	private explicit impl core::default::Default for d2b_core_controller::authority::quota::Usage negative=false generic-params=0 where-predicates=0
 d2b_core_controller::d2b_core_controller::authority_persistence::AuthorityRecoveryCoordinator	struct	core::core::fmt::Debug	trait	impl core::fmt::Debug for d2b_core_controller::authority_persistence::AuthorityRecoveryCoordinator
-d2b_core_controller::d2b_core_controller::authority_persistence::RedbAuthorityPersistence	struct	core::core::fmt::Debug	trait	impl core::fmt::Debug for d2b_core_controller::authority_persistence::RedbAuthorityPersistence
-d2b_core_controller::d2b_core_controller::authority_persistence::RedbAuthorityPersistence	struct	d2b_core_controller::d2b_core_controller::authority_persistence::AuthorityPersistence	trait	private explicit impl d2b_core_controller::authority_persistence::AuthorityPersistence for d2b_core_controller::authority_persistence::RedbAuthorityPersistence negative=false generic-params=0 where-predicates=0
-d2b_core_controller::d2b_core_controller::authority_persistence::RedbAuthorityPersistence	struct	d2b_core_controller::d2b_core_controller::authority_persistence::AuthorityRecoveryProvenance	trait	private explicit impl d2b_core_controller::authority_persistence::AuthorityRecoveryProvenance for d2b_core_controller::authority_persistence::RedbAuthorityPersistence negative=false generic-params=0 where-predicates=0
-d2b_core_controller::d2b_core_controller::authority_persistence::RedbAuthorityPersistence	struct	d2b_core_controller::d2b_core_controller::authority_persistence::sealed::RecoveryProvenance	trait	private explicit impl d2b_core_controller::authority_persistence::sealed::RecoveryProvenance for d2b_core_controller::authority_persistence::RedbAuthorityPersistence negative=false generic-params=0 where-predicates=0
 d2b_core_controller::d2b_core_controller::authz::AuthorizationHandler	struct	core::core::default::Default	trait	impl core::default::Default for d2b_core_controller::authz::AuthorizationHandler
 d2b_core_controller::d2b_core_controller::authz::AuthorizationHandler	struct	core::core::fmt::Debug	trait	impl core::fmt::Debug for d2b_core_controller::authz::AuthorizationHandler
 d2b_core_controller::d2b_core_controller::authz::CoreAuthorizationSubject	struct	core::core::fmt::Debug	trait	impl core::fmt::Debug for d2b_core_controller::authz::CoreAuthorizationSubject<'_>
@@ -1773,6 +1769,9 @@ d2bd::d2bd::DaemonUsbipStopCleanupExecutor	struct	d2bd::d2bd::usbip_reconcile_st
 d2bd::d2bd::ServerState	struct	core::core::clone::Clone	trait	private explicit impl core::clone::Clone for d2bd::ServerState negative=false generic-params=0 where-predicates=0
 d2bd::d2bd::ServerState	struct	core::core::fmt::Debug	trait	private explicit impl core::fmt::Debug for d2bd::ServerState negative=false generic-params=0 where-predicates=0
 d2bd::d2bd::VmStartRunner	struct	d2bd::d2bd::supervisor::dag::NodeRunner	trait	private explicit impl d2bd::supervisor::dag::NodeRunner for d2bd::VmStartRunner negative=false generic-params=0 where-predicates=0
+d2bd::d2bd::authority_persistence::RedbAuthorityPersistence	struct	core::core::fmt::Debug	trait	private explicit impl core::fmt::Debug for d2bd::authority_persistence::RedbAuthorityPersistence negative=false generic-params=0 where-predicates=0
+d2bd::d2bd::authority_persistence::RedbAuthorityPersistence	struct	d2b_core_controller::d2b_core_controller::authority_persistence::AuthorityPersistence	trait	private explicit impl d2b_core_controller::authority_persistence::AuthorityPersistence for d2bd::authority_persistence::RedbAuthorityPersistence negative=false generic-params=0 where-predicates=0
+d2bd::d2bd::authority_persistence::RedbAuthorityPersistence	struct	d2b_core_controller::d2b_core_controller::authority_persistence::AuthorityRecoveryProvenance	trait	private explicit impl d2b_core_controller::authority_persistence::AuthorityRecoveryProvenance for d2bd::authority_persistence::RedbAuthorityPersistence negative=false generic-params=0 where-predicates=0
 d2bd::d2bd::provider_effects::GuestLifecycleRequest	struct	core::core::clone::Clone	trait	impl core::clone::Clone for d2bd::provider_effects::GuestLifecycleRequest
 d2bd::d2bd::provider_effects::GuestLifecycleRequest	struct	core::core::cmp::Eq	trait	impl core::cmp::Eq for d2bd::provider_effects::GuestLifecycleRequest
 d2bd::d2bd::provider_effects::GuestLifecycleRequest	struct	core::core::cmp::PartialEq	trait	impl core::cmp::PartialEq for d2bd::provider_effects::GuestLifecycleRequest
