+++ tests/golden/api-surface/hidden-public-api.txt
@@ -1,4 +1,14 @@
 d2b_core::d2b_core::for_tests	function	function
+d2b_core_controller::d2b_core_controller::authority::claim_digest	function	function
+d2b_core_controller::d2b_core_controller::generation	function	function
+d2b_core_controller::d2b_core_controller::host_uid	function	function
+d2b_core_controller::d2b_core_controller::identity_digest	function	function
+d2b_core_controller::d2b_core_controller::nonce	function	function
+d2b_core_controller::d2b_core_controller::operation_id	function	function
+d2b_core_controller::d2b_core_controller::owner_proof	function	function
+d2b_core_controller::d2b_core_controller::resource_ref	function	function
+d2b_core_controller::d2b_core_controller::resource_uid	function	function
+d2b_core_controller::d2b_core_controller::store_binding_digest	function	function
 d2b_session::d2b_session::revoke_generation_writes	function	function
 d2b_session::d2b_session::transport::OwnedTransport::begin_write_batch	function	function
 d2b_session::d2b_session::transport::OwnedTransport::take_write_batch	function	function
