+++ tests/golden/api-surface/input-fingerprint.txt
@@ -1,2 +1,2 @@
 version=1
-sha256=0ce81548a0caba48082e20c9c55c82922208822054ab8cd389f3607f04cf2c73
+sha256=cdf595c0e4f79e41d02449980ce1a1f6635978949939b98258cc576686795187
