+++ tests/golden/api-surface/workspace-metadata.json
@@ -53,14 +53,14 @@
       "private_census": {
         "external_crates": 131,
         "index_items": 9102,
-        "path_items": 14158
+        "path_items": 14155
       },
       "private_hidden_items": 146,
       "private_json_file": "d2b_bus.json",
       "public_census": {
         "external_crates": 131,
         "index_items": 4205,
-        "path_items": 13987
+        "path_items": 13984
       },
       "public_json_file": "d2b_bus.json"
     },
@@ -147,16 +147,16 @@
     {
       "crate_name": "d2b_core_controller",
       "private_census": {
-        "external_crates": 70,
-        "index_items": 17867,
-        "path_items": 12411
+        "external_crates": 67,
+        "index_items": 17907,
+        "path_items": 12158
       },
-      "private_hidden_items": 369,
+      "private_hidden_items": 381,
       "private_json_file": "d2b_core_controller.json",
       "public_census": {
-        "external_crates": 70,
-        "index_items": 9916,
-        "path_items": 12313
+        "external_crates": 67,
+        "index_items": 9933,
+        "path_items": 12061
       },
       "public_json_file": "d2b_core_controller.json"
     },
@@ -467,16 +467,16 @@
     {
       "crate_name": "d2b_provider_device_tpm",
       "private_census": {
-        "external_crates": 71,
+        "external_crates": 68,
         "index_items": 1392,
-        "path_items": 11747
+        "path_items": 11492
       },
       "private_hidden_items": 34,
       "private_json_file": "d2b_provider_device_tpm.json",
       "public_census": {
-        "external_crates": 71,
+        "external_crates": 68,
         "index_items": 811,
-        "path_items": 11697
+        "path_items": 11442
       },
       "public_json_file": "d2b_provider_device_tpm.json"
     },
@@ -549,14 +549,14 @@
       "private_census": {
         "external_crates": 127,
         "index_items": 1257,
-        "path_items": 13432
+        "path_items": 13429
       },
       "private_hidden_items": 23,
       "private_json_file": "d2b_provider_observability_otel.json",
       "public_census": {
         "external_crates": 127,
         "index_items": 712,
-        "path_items": 13367
+        "path_items": 13364
       },
       "public_json_file": "d2b_provider_observability_otel.json"
     },
@@ -709,14 +709,14 @@
       "private_census": {
         "external_crates": 128,
         "index_items": 580,
-        "path_items": 13415
+        "path_items": 13412
       },
       "private_hidden_items": 8,
       "private_json_file": "d2b_provider_system_systemd.json",
       "public_census": {
         "external_crates": 128,
         "index_items": 331,
-        "path_items": 13355
+        "path_items": 13352
       },
       "public_json_file": "d2b_provider_system_systemd.json"
     },
@@ -725,14 +725,14 @@
       "private_census": {
         "external_crates": 126,
         "index_items": 2670,
-        "path_items": 13442
+        "path_items": 13439
       },
       "private_hidden_items": 37,
       "private_json_file": "d2b_provider_toolkit.json",
       "public_census": {
         "external_crates": 126,
         "index_items": 1516,
-        "path_items": 13360
+        "path_items": 13357
       },
       "public_json_file": "d2b_provider_toolkit.json"
     },
@@ -901,14 +901,14 @@
       "private_census": {
         "external_crates": 101,
         "index_items": 4560,
-        "path_items": 13013
+        "path_items": 13010
       },
       "private_hidden_items": 64,
       "private_json_file": "d2b_resource_api.json",
       "public_census": {
         "external_crates": 101,
         "index_items": 1989,
-        "path_items": 12848
+        "path_items": 12845
       },
       "public_json_file": "d2b_resource_api.json"
     },
@@ -981,14 +981,14 @@
       "private_census": {
         "external_crates": 123,
         "index_items": 4555,
-        "path_items": 13504
+        "path_items": 13501
       },
       "private_hidden_items": 38,
       "private_json_file": "d2b_session.json",
       "public_census": {
         "external_crates": 123,
         "index_items": 1816,
-        "path_items": 13278
+        "path_items": 13275
       },
       "public_json_file": "d2b_session.json"
     },
@@ -997,14 +997,14 @@
       "private_census": {
         "external_crates": 130,
         "index_items": 3092,
-        "path_items": 13695
+        "path_items": 13692
       },
       "private_hidden_items": 35,
       "private_json_file": "d2b_session_unix.json",
       "public_census": {
         "external_crates": 130,
         "index_items": 1335,
-        "path_items": 13564
+        "path_items": 13561
       },
       "public_json_file": "d2b_session_unix.json"
     },
@@ -1076,15 +1076,15 @@
       "crate_name": "d2bd",
       "private_census": {
         "external_crates": 241,
-        "index_items": 41217,
-        "path_items": 21241
+        "index_items": 41306,
+        "path_items": 21246
       },
       "private_hidden_items": 356,
       "private_json_file": "d2bd.json",
       "public_census": {
         "external_crates": 241,
         "index_items": 14388,
-        "path_items": 20015
+        "path_items": 20017
       },
       "public_json_file": "d2bd.json"
     }
